// src/common/enums/pay-method-type.ts
export enum PayMethodType {
  Samsung = 'samsung',
  Card = 'card',
  Trans = 'trans',
  Vbank = 'vbank',
  Phone = 'phone',
  Cultureland = 'cultureland',
  Smartculture = 'smartculture',
  Booknlife = 'booknlife',
  Happymoney = 'happymoney',
  Point = 'point',
  Ssgpay = 'ssgpay',
  Lpay = 'lpay',
  Payco = 'payco',
  Kakaopay = 'kakaopay',
  Tosspay = 'tosspay',
  Naverpay = 'naverpay',
}